requirements.txt
Flask
gunicorn
requests
Werkzeug
feedparser
mistune
python-dateutil
pytz